# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Sirens Stream Agency API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tutorials/{id}:
    get:
      operationId: getTutorial
      tags: [tutorials]
      summary: Get a single tutorial
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tutorial detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tutorial"
        "404":
          description: Not found
...
components:
  schemas:
...
    Tutorial:
      type: object
      required:
        - id
        - title
        - description
        - category
        - duration
        - level
        - imageUrl
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [setup, monetization, growth, technical, branding]
        duration:
          type: string
        level:
          type: string
          enum: [beginner, intermediate, advanced]
        imageUrl:
          type: string
        videoUrl:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
...